Split SSL ECH tests
