+++ ssl/CMakeLists.txt
@@ -87,6 +87,7 @@ if(BUILD_TESTING)
     ssl_test_common.cc
     ssl_misc_test.cc
     ssl_version_test.cc
+    ssl_ech_test.cc
   )
 
   target_link_libraries(${SSL_TEST_EXEC} boringssl_gtest_main ssl)
